refactored addPetForm page to dialog
